Normalizzazione fitness e risoluzione problemi
- Normalizzato il risultato della funzione di fitness in MapBrain e cambiati i tipi dei parametri dell'algoritmo genetico (int -> float)
- Implementato il metodo RebuildGrid in CandidateMap per ricostruire la griglia in base ai dati ottenuti dall'algoritmo genetico

+++ Genetic Map Generation/Scripts/GeneticAlgorithmConfiguration.cs
@@ -6,9 +6,9 @@ public struct GeneticAlgorithmConfiguration
 	public int mutationRate;
 	public int fitnessCornerMin;
 	public int fitnessCornerMax;
-	public int fitnessCornerWeight;
-	public int fitnessObstacleWeight;
-	public int fitnessPathWeight;
+	public float fitnessCornerWeight;
+	public float fitnessObstacleWeight;
+	public float fitnessPathWeight;
 	public int mapWidth;
 	public int mapHeight;
 	public int numberOfPieces;
