+++ Genetic Map Generation/Scripts/GeneticAlgorithmData.cs
@@ -2,8 +2,8 @@ using System;
 
 public struct GeneticAlgorithmData
 {
-	public int[] fitnessArray;
-    public int bestMapFitness;
+	public float[] fitnessArray;
+    public float bestMapFitness;
     public int pathLenght;
     public int numberOfCorners;
     public int numberOfObstacles;
